Significant rewrite of the tokenizer machinery.
Taking a compile-time approach to building the state transition table.
The table no longer features O(1) lookup using only a From state and a
CurrentChar. This was to support, very large ranges in Unicode that
would expand to ~4gib just for a single pattern match if using the
HashTable approach. This means to lookup a transition, an O(N) lookup
is now made using the State and CurrentChar, but instead of direct
char lookup, it is tested to see whether it resides in a range and
not in a range, respectively. Although this operation is very cheap,
it also means the lookup will scale with table size.